# internal/handler/endpoint/endpoint.go
package endpoint

// Endpoint for registering a user
const Register = "/api/v1/register"
const Login = "/api/v1/login"
const CreateServiceUser = "/api/v1/service/register"
const LoginServiceUser = "/api/v1/service/login"
const TokenRefresh = "/api/v1/token/refresh"
const UpdatePassword = "/api/v1/user/password/update"